Renamed OpenModDbContext to OpenModDbContextBase and removed MySQL specific features
