+++ framework/OpenMod.EntityFrameworkCore/Extensions/EntityFrameworkCoreContainerBuilderExtensions.cs
@@ -9,22 +9,22 @@ namespace OpenMod.EntityFrameworkCore.Extensions
 {
     public static class EntityFrameworkCoreContainerBuilderExtensions
     {
-        public static ContainerBuilder AddDbContext<T>(this ContainerBuilder containerBuilder) where T : OpenModDbContext<T>
+        public static ContainerBuilder AddDbContext<T>(this ContainerBuilder containerBuilder) where T : OpenModDbContextBase<T>
         {
             return AddDbContextInternal(containerBuilder, typeof(T), null, null);
         }
 
-        public static ContainerBuilder AddDbContext<T>(this ContainerBuilder containerBuilder, ServiceLifetime serviceLifetime) where T : OpenModDbContext<T>
+        public static ContainerBuilder AddDbContext<T>(this ContainerBuilder containerBuilder, ServiceLifetime serviceLifetime) where T : OpenModDbContextBase<T>
         {
             return AddDbContextInternal(containerBuilder, typeof(T), null, serviceLifetime);
         }
 
-        public static ContainerBuilder AddDbContext<T>(this ContainerBuilder containerBuilder, Action<DbContextOptionsBuilder>? optionsBuilder) where T : OpenModDbContext<T>
+        public static ContainerBuilder AddDbContext<T>(this ContainerBuilder containerBuilder, Action<DbContextOptionsBuilder>? optionsBuilder) where T : OpenModDbContextBase<T>
         {
             return AddDbContextInternal(containerBuilder, typeof(T), optionsBuilder, null);
         }
 
-        public static ContainerBuilder AddDbContext<T>(this ContainerBuilder containerBuilder, Action<DbContextOptionsBuilder>? optionsBuilder, ServiceLifetime serviceLifetime) where T : OpenModDbContext<T>
+        public static ContainerBuilder AddDbContext<T>(this ContainerBuilder containerBuilder, Action<DbContextOptionsBuilder>? optionsBuilder, ServiceLifetime serviceLifetime) where T : OpenModDbContextBase<T>
         {
             return AddDbContextInternal(containerBuilder, typeof(T), optionsBuilder, serviceLifetime);
         }
@@ -47,17 +47,6 @@ namespace OpenMod.EntityFrameworkCore.Extensions
         public static ContainerBuilder AddDbContext(this ContainerBuilder containerBuilder, Type dbContextType, Action<DbContextOptionsBuilder>? optionsBuilder, ServiceLifetime serviceLifetime)
         {
             return AddDbContextInternal(containerBuilder, dbContextType, optionsBuilder, serviceLifetime);
-        }
-
-        public static ContainerBuilder AddEntityFrameworkCoreMySql(this ContainerBuilder containerBuilder)
-        {
-            containerBuilder.RegisterType<ConfigurationBasedConnectionStringAccessor>()
-                .As<ConfigurationBasedConnectionStringAccessor>()
-                .As<IConnectionStringAccessor>()
-                .OwnedByLifetimeScope()
-                .InstancePerDependency();
-
-            return containerBuilder;
         }
 
         private static ContainerBuilder AddDbContextInternal(this ContainerBuilder containerBuilder,
